:white_check_mark: add and update tests
