+++ quant_candles/tests/controllers/test_trades.py
@@ -1,0 +1,155 @@
+from datetime import datetime, timezone
+from typing import List, Tuple
+from unittest.mock import patch
+
+import pandas as pd
+import time_machine
+
+from quant_candles.constants import Frequency
+from quant_candles.controllers import TradeDataIterator
+from quant_candles.models import TradeData
+
+from ..base import BaseSymbolTest
+
+
+@time_machine.travel(datetime(2009, 1, 3))
+class TradeDataIteratorTest(BaseSymbolTest):
+    def setUp(self):
+        super().setUp()
+        self.one_minute = pd.Timedelta("1t")
+        self.timestamp_to = self.timestamp_from + (self.one_minute * 5)
+        self.symbol = self.get_symbol()
+
+    def get_values(self, retry: bool = False) -> List[Tuple[datetime, datetime]]:
+        """Get values."""
+        return [
+            value
+            for value in TradeDataIterator(self.symbol).iter_all(
+                timestamp_from=self.timestamp_from,
+                timestamp_to=self.timestamp_to,
+                retry=retry,
+            )
+        ]
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 3).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_no_results(self, mock_get_max_timestamp_to):
+        """No results."""
+        values = self.get_values()
+        self.assertEqual(len(values), 0)
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 4).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_head(self, mock_get_max_timestamp_to):
+        """First is OK."""
+        TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_from,
+            frequency=Frequency.MINUTE,
+            ok=True,
+        )
+        values = self.get_values()
+        self.assertEqual(len(values), 1)
+        self.assertEqual(values[0][0], self.timestamp_from + self.one_minute)
+        self.assertEqual(values[-1][1], self.timestamp_to)
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 4).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_one_ok(self, mock_get_max_timestamp_to):
+        """Second is OK."""
+        obj = TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_from + self.one_minute,
+            frequency=Frequency.MINUTE,
+            ok=True,
+        )
+        values = self.get_values()
+        self.assertEqual(len(values), 2)
+        self.assertEqual(values[0][0], obj.timestamp + self.one_minute)
+        self.assertEqual(values[0][1], self.timestamp_to)
+        self.assertEqual(values[-1][0], self.timestamp_from)
+        self.assertEqual(values[-1][1], obj.timestamp)
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 4).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_two_ok(self, mock_get_max_timestamp_to):
+        """Second and fourth are OK."""
+        obj_one = TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_from + self.one_minute,
+            frequency=Frequency.MINUTE,
+            ok=True,
+        )
+        obj_two = TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_from + (self.one_minute * 3),
+            frequency=Frequency.MINUTE,
+            ok=True,
+        )
+        values = self.get_values()
+        self.assertEqual(len(values), 3)
+        self.assertEqual(values[0][0], self.timestamp_to - self.one_minute)
+        self.assertEqual(values[0][1], self.timestamp_to)
+        self.assertEqual(values[1][0], obj_one.timestamp + self.one_minute)
+        self.assertEqual(values[1][1], obj_two.timestamp)
+        self.assertEqual(values[-1][0], self.timestamp_from)
+        self.assertEqual(values[-1][1], obj_one.timestamp)
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 4).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_tail(self, mock_get_max_timestamp_to):
+        """Last is OK."""
+        TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_to - self.one_minute,
+            frequency=Frequency.MINUTE,
+            ok=True,
+        )
+        values = self.get_values()
+        self.assertEqual(len(values), 1)
+        self.assertEqual(values[0][0], self.timestamp_from)
+        self.assertEqual(values[0][1], self.timestamp_to - self.one_minute)
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 4).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_retry_and_one_not_ok(self, mock_get_max_timestamp_to):
+        """One is not OK."""
+        TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_from,
+            frequency=Frequency.MINUTE,
+            ok=False,
+        )
+        values = self.get_values(retry=True)
+        self.assertEqual(len(values), 1)
+        self.assertEqual(values[0][0], self.timestamp_from)
+        self.assertEqual(values[-1][1], self.timestamp_to)
+
+    @patch(
+        "quant_candles.controllers.iterators.TradeDataIterator.get_max_timestamp_to",
+        return_value=datetime(2009, 1, 4).replace(tzinfo=timezone.utc),
+    )
+    def test_iter_all_with_retry_and_one_missing(self, mock_get_max_timestamp_to):
+        """One is missing."""
+        TradeData.objects.create(
+            symbol=self.symbol,
+            timestamp=self.timestamp_from,
+            frequency=Frequency.MINUTE,
+            ok=None,
+        )
+        values = self.get_values(retry=True)
+        self.assertEqual(len(values), 1)
+        self.assertEqual(values[0][0], self.timestamp_from + self.one_minute)
+        self.assertEqual(values[-1][1], self.timestamp_to)
